requirements.txt
urllib3==1.26.1
python_telegram_bot==13.1